make header file for menu 2(View Cooking Process)
